refactor: DTO 필드명 수정 및 테스트 경로 정리 - PopularBookDto : 필드명을 thumbnailFileName → thumbnailUrl 로 재변경 - batch 하위 경로 분리와 분리에 맞춰 테스트 경로 변경
