+++ src/main/java/com/part3/team07/sb01deokhugamteam07/dto/book/PopularBookDto.java
@@ -9,7 +9,7 @@ public record PopularBookDto(
     UUID bookId,
     String title,
     String author,
-    String thumbnailFileName,
+    String thumbnailUrl,
     Period period,
     int rank,
     double score,
